fix: specializations and deduction guides are listed in order
In both lists, all elements take the name of the primary, so comparing
names never decided anything, and the order fell through to the symbol
IDs, which are hashes of the declaration.

They now compare on what is there in them beyond that shared name: what
a specialization was specialized with, and what a guide deduces, with
the types a guide takes deciding between guides that deduce the same
thing.

Fixes #1250.
Fixes #1260.
