+++ tests/golden/fixtures/config/sort/unordered.xml
@@ -74,7 +74,7 @@
       </type>
     </params>
   </template>
-  <specializations>jc6pZFjd9Tdq7AfNKC8Gpu96PGb 3yvgGWnjAPNP6H1PLANzkvLTLW9z</specializations>
+  <specializations>3yvgGWnjAPNP6H1PLANzkvLTLW9z jc6pZFjd9Tdq7AfNKC8Gpu96PGb</specializations>
 </record>
 <record>
   <name>C</name>
