+++ tests/golden/fixtures/javadoc/copydoc/template-arguments.xml
@@ -771,7 +771,7 @@
       </constant>
     </params>
   </template>
-  <specializations>3j4K5kCQDE35YBWkduVPJ1yGQowL 4TQvKQDPskqVPACGDhax7XqREUGx</specializations>
+  <specializations>4TQvKQDPskqVPACGDhax7XqREUGx 3j4K5kCQDE35YBWkduVPJ1yGQowL</specializations>
 </record>
 <record>
   <name>C</name>
