+++ tests/golden/fixtures/symbols/function/spec-mem-implicit-instantiation.xml
@@ -40,7 +40,7 @@
       <functions>2cMeQjuQcpTPhbY6f3vK8Un6jDdo</functions>
     </public>
   </interface>
-  <specializations>1YWrSpfV88zbfQ3PBCGBoYZMGKW ARF3LuXnz2naUZPeNsVbYJsWAoc 4Z2a1HFinPWXq4fVNWV3jnmaLCeM</specializations>
+  <specializations>4Z2a1HFinPWXq4fVNWV3jnmaLCeM 1YWrSpfV88zbfQ3PBCGBoYZMGKW ARF3LuXnz2naUZPeNsVbYJsWAoc</specializations>
 </record>
 <record>
   <name>B</name>
