+++ tests/golden/fixtures/symbols/guide/explicit-deduct-guide.xml
@@ -45,7 +45,7 @@
       </constant>
     </params>
   </template>
-  <deductionGuides>3HabhwAH81oNwG5BP2g96XC5TBK 2MXo6mBWH4UeLLE5YkZKZEN9T4BU 3YHV7JqPJcNccg1srPhakFGF9cQw 41woAuDtuyrTsDGUHTUkVrj9Pu3Q</deductionGuides>
+  <deductionGuides>41woAuDtuyrTsDGUHTUkVrj9Pu3Q 3HabhwAH81oNwG5BP2g96XC5TBK 3YHV7JqPJcNccg1srPhakFGF9cQw 2MXo6mBWH4UeLLE5YkZKZEN9T4BU</deductionGuides>
 </record>
 <guide>
   <name>X</name>
