+++ tests/golden/fixtures/symbols/record/class-template-specializations-2.xml
@@ -34,7 +34,7 @@
       </type>
     </params>
   </template>
-  <specializations>cFbuJ5pS7LuCqTFzp5U2gXyJvog ksfLe6aja8PH8zMMK1QyzdX3JH7 2XgfAs1uQHGKjrRevxW93SmXk77p</specializations>
+  <specializations>2XgfAs1uQHGKjrRevxW93SmXk77p cFbuJ5pS7LuCqTFzp5U2gXyJvog ksfLe6aja8PH8zMMK1QyzdX3JH7</specializations>
 </record>
 <record>
   <name>A</name>
@@ -754,7 +754,7 @@
       </type>
     </params>
   </template>
-  <specializations>QRmd63tDYPdAzhpKMP8nPkoZ59V 2FdnBZwJ7A3iUaNLhzGkiPVfm5a6 3fGmtG1N8Y9fbQ8WWkuvr55oC5fF 46XBVeLuan9q8y7t9krr5yd2D6jc</specializations>
+  <specializations>3fGmtG1N8Y9fbQ8WWkuvr55oC5fF 46XBVeLuan9q8y7t9krr5yd2D6jc 2FdnBZwJ7A3iUaNLhzGkiPVfm5a6 QRmd63tDYPdAzhpKMP8nPkoZ59V</specializations>
 </record>
 <record>
   <name>B</name>
