+++ tests/golden/fixtures/symbols/record/class-template-specializations-3.xml
@@ -39,7 +39,7 @@
       <records>2dEvF5AtxD2tmCfHnQo2Q4zRVYij 3XBBL39AxNcm4rkZynuiuyyFjWH</records>
     </public>
   </interface>
-  <specializations>ARF3LuXnz2naUZPeNsVbYJsWAoc LfSCdYEqubuTVzw8qrDfpmQY7Zv 2hWNBnvdRML25tDFQ6A2jmhLQv6K 3qqXmqMRwZDgqmj4UNGSCBe4h9Re</specializations>
+  <specializations>LfSCdYEqubuTVzw8qrDfpmQY7Zv 3qqXmqMRwZDgqmj4UNGSCBe4h9Re ARF3LuXnz2naUZPeNsVbYJsWAoc 2hWNBnvdRML25tDFQ6A2jmhLQv6K</specializations>
 </record>
 <record>
   <name>B</name>
@@ -72,7 +72,7 @@
       <records>2YDKvd48C3QScNe6TEcg2DJKzvGU 39AyAekzAMFUtzGKtJ2ZiSm5VQYx 4NahHsYGaefEDqtfVyjeUTzFZvDE</records>
     </public>
   </interface>
-  <specializations>2xHHUcPMhZck4AozMtVgSnADL5e 3XBBL39AxNcm4rkZynuiuyyFjWH sUSbDg9t8EBktmnpTsGUf6Mjy2B 2LSBQtnc9YVtk6GvWuGyRRkY9UBr 2k7G7bCdrTFeAx4Fo9LPyjQRpe83 2o6NLAjXd7GxdsjTe1vreE5vQ811</specializations>
+  <specializations>3XBBL39AxNcm4rkZynuiuyyFjWH sUSbDg9t8EBktmnpTsGUf6Mjy2B 2k7G7bCdrTFeAx4Fo9LPyjQRpe83 2o6NLAjXd7GxdsjTe1vreE5vQ811 2LSBQtnc9YVtk6GvWuGyRRkY9UBr 2xHHUcPMhZck4AozMtVgSnADL5e</specializations>
 </record>
 <record>
   <name>C</name>
@@ -572,7 +572,7 @@
       <records>CoggmorQNBUZutUzrTWG39eZ2Gi 3cJvwJcpr99i7ANoLDiRPDCpZYh 3fHTDLNMmrCEZQpLkPJD8ZNbZ1W4</records>
     </public>
   </interface>
-  <specializations>2zXjjixuaNK8m6HKyua7kmdES4m1 46M27E53RAM4YWyS9dxfN5twW9LQ</specializations>
+  <specializations>46M27E53RAM4YWyS9dxfN5twW9LQ 2zXjjixuaNK8m6HKyua7kmdES4m1</specializations>
 </record>
 <record>
   <name>C</name>
