+++ tests/golden/fixtures/symbols/record/specialization-order.yml
@@ -1,0 +1,1 @@
+generator: adoc
