+++ tests/golden/fixtures/symbols/record/template-specialization-inheritance.xml
@@ -342,7 +342,7 @@
       <records>41eimfG9ZxfyKecXgT1Pu8aYc7Fv</records>
     </public>
   </interface>
-  <specializations>2cVSEcoiC3S52ZWJahZas135jQxo 2m73ELeQ8FvdfiQ1Bjum6GYoqtww 3BWnweXEGCZThEU86yPvT6zBYgcJ 4FR11bfr43aruqgBV4G2pESSS68N</specializations>
+  <specializations>3BWnweXEGCZThEU86yPvT6zBYgcJ 2cVSEcoiC3S52ZWJahZas135jQxo 4FR11bfr43aruqgBV4G2pESSS68N 2m73ELeQ8FvdfiQ1Bjum6GYoqtww</specializations>
 </record>
 <record>
   <name>S1</name>
