+++ tests/golden/fixtures/symbols/typedef/decay-to-primary.xml
@@ -150,7 +150,7 @@
       <typedefs>2T6kf4JHdfvP81ik6iGWmbMCwm9F 419VWzEKhdU748cQojDXRFJS5cQu</typedefs>
     </public>
   </interface>
-  <specializations>5NPyEt9hdXVw7hsVatqZvuF1N49 3Bht2tVCEYBFkQZKdLcUDbpJSEWz</specializations>
+  <specializations>3Bht2tVCEYBFkQZKdLcUDbpJSEWz 5NPyEt9hdXVw7hsVatqZvuF1N49</specializations>
 </record>
 <typedef>
   <name>M0</name>
